Refactored app structure by introducing custom hooks (`useCart`, `useOrderForm`, `useCheckout`), extracted Sheets API to `sheetsApi.js`, and simplified `App.js` component.
